+++ packages/terra-core-docs/CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Unreleased
 
 * Updated
-  * Changed status view title font size as per OpenSource Theme
+  * Changed status view title font size as per OpenSource Theme.
 
 ## 1.27.0 - (May 9, 2023)