+++ packages/terra-core-docs/CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Unreleased
 
 * Added
- * Added terra-scroll A11y tests
+ * Added terra-scroll A11y tests.
 
 ## 1.27.0 - (May 9, 2023)